Cast p_ref to the network dtype
Production runs float64 momenta (data.momentum_float64: true) through
float32 nets. lloca 2.0 folds gamma = (L p_ref)^0 / m_ref into the frames,
so a float64 p_ref silently upcast the whole q/k/v transport to float64
(verified on LLoCaAttention: frames_out came back float64 while the output
stayed float32). The frames were already cast to the network dtype; p_ref
now is too, in the tagging base wrapper and the amplitude wrapper.

tests/experiments/test_preserve_variance.py::test_p_ref_keeps_the_network_dtype
pins the production dtype combination for three model families.

+++ experiments/tagging/wrappers.py
@@ -291,7 +291,10 @@ class TaggerWrapper(nn.Module):
             reduce="sum",
             dim_size=B,
         )  # (B, 4)
-        self._p_ref = jet_global
+        # network dtype like the frames below: a float64 p_ref with float32 frames would
+        # silently upcast lloca's whole q/k/v transport to float64 (gamma ~ 1..1e2 needs no
+        # extra precision)
+        self._p_ref = jet_global.to(scalars_nospurions.dtype)
         jet_nospurions = jet_global.index_select(0, batch_nospurions)
         jet_local_nospurions = self.trafo_fourmomenta(jet_nospurions, frames_nospurions)
         local_tagging_features_nospurions = get_tagging_features(
